Color links as common type
